fix init opts handling in connection (#2)
This pull request fixes `init_opts` handling when the given options are not a list of configs or a list of URIs

We save the first given options to the connection so that we can recover the initial state in case of reconnection and not lose fallback. 

When we reconnect we want to remove the current connection from the `init_opts` and keep this as a fallback configuration list. This led to a bug because we were trying to `List.delete` on something that wasn't containing the right data.